R3.5.1 with Conda
This converts our rocket based Rstudio to one that runs inside a conda
environment. This allows us to cut down on the packages we install into the base
debian environment.

The upside of this is:

1. Users can use `conda` instead of `packrat` to manage their dependencies
2. Users can install OS level dependencies e.g. libjpeg or whatever
3. If the user wants to, they can use conda to change their version of R

How does it work?
-----------------

The base image include a `base` docker environment with the packages required to
run Rstudio and compile common packages through packrat (meaning this is
backwards compatible, they don't need to switch to `conda` from `packrat`
immediately.

Won't the `RStudio` environment be reset on every boot?
-------------------------------------------------------

To avoid this the `start.sh` script clones the base environment into the user's
`~/.conda/envs/rstudio` directory so it is local to them and lives on their NFS share.
This is only done once. Any packages they install into the rstudio `conda`
environment will persist across reboots.

Will cloning the base conda environment be slow?
------------------------------------------------

Yes, that's why the `start.sh` script has been refactored to allow only running
the init steps in an init container by passing an extra argument to the
`start.sh` script.
It remains backwards compatible and will still do the whole shebang if no arg is
passed to it.